Move MethodUsages to a separate file to minimize expect diffs.
Previously, adding a line to `example.Methods` would introduce large
diffs in metac.expect since all call-sites in `MethodsUsages` would
shift by a couple line numbers. This commit moves `MethodsUsages` to a
separate file so that new definitions in `example.Methods` introduce
minimal diffs.
