+++ tests/jvm/src/test/resources/metac.index
@@ -112,7 +112,7 @@ example.B# => semanticdb/integration/src/main/scala/example/Overrides.scala.sema
 example.EmptyObject. => semanticdb/integration/src/main/scala/example/EmptyObject.scala.semanticdb
 example.Example. => semanticdb/integration/src/main/scala/example/Example.scala.semanticdb
 example.MetacJava# => semanticdb/integration/src/main/scala/example/MetacJava.scala.semanticdb
-example.MethodUsages# => semanticdb/integration/src/main/scala/example/Methods.scala.semanticdb
+example.MethodUsages# => semanticdb/integration/src/main/scala/example/MethodUsages.scala.semanticdb
 example.Methods# => semanticdb/integration/src/main/scala/example/Methods.scala.semanticdb
 example.Synthetic# => semanticdb/integration/src/main/scala/example/Synthetic.scala.semanticdb
 example.ValUsages. => semanticdb/integration/src/main/scala/example/Vals.scala.semanticdb
